Prettified code and fixed typo
