+++ statementsPdf/generate.go
@@ -75,7 +75,7 @@ func (p cpdf) renderStatementsTable(data [][]string, headers []string) {
 			if count > len(data) {
 				count = 1
 			}
-			cell.str = strings.Join(data[(len(data) -1)-row][col:col+1], " ")
+			cell.str = strings.Join(data[(len(data)-1)-row][col:col+1], " ")
 			cell.list = p.Fpdf.SplitLines([]byte(cell.str), colWd-cellGap-cellGap)
 			cell.ht = float64(len(cell.list)) * lineHt
 			if cell.ht > maxHt {
